fixed variable name

+++ ADXL/ADXL345.cpp
@@ -1539,7 +1539,7 @@ void ADXL345::getAcceleration(int16_t* x, int16_t* y, int16_t* z) {
 	// *y = i2c.readWord(ADXL345_RA_DATAY1, ADXL345_RA_DATAY0);
 	// *z = i2c.readWord(ADXL345_RA_DATAZ1, ADXL345_RA_DATAZ0);
 
-	uint8_t array = i2c.readStream(ADXL345_RA_DATAX0,ADXL345_RA_DATAX1,
+	uint16_t array = i2c.readStream(ADXL345_RA_DATAX0,ADXL345_RA_DATAX1,
 		ADXL345_RA_DATAY0,ADXL345_RA_DATAY1,ADXL345_RA_DATAZ0,ADXL345_RA_DATAZ1);
 
 	*x = array[0];
